Run request tests from smithy-aws-protocol-tests
Instead of code generating tests, we load the model up at test
initialization time and use JUnit5's test templates to generate a test
for every httpRequestTest in the model. We also spin up a netty server
for each service in the model on its own port. Then we make a request,
capture the deserialized object that the operation is invoked with, and
compare it structurally to the params in the modeled httpRequestTest.