Make JSON codecs pluggable, use Jackson
json-iter has some interesting behavior around handling values that
start with n (specifically, it assumes it's encountered null and just
skips the next three bytes). So that we can later test JSON
implementations against each other, or plug in different parsers (such
as simd-json on JDK21+ servers), we load different providers at startup
and allow the user to configure a preferred provider with a system
property.